+++ Render/image_define/Levels/Water.java
@@ -4,7 +4,7 @@ import javafx.scene.image.Image;
 
 public class Water extends Block{
     Image skin2;
-    public Water(long x,long y,long width,long height){
+    public Water(double x,double y,double width,double height){
         super( x, y, width, height,"null");
         //skin = new Image(path_to)
         //skin2 = new Image(path_to)
